rename clientsview
